requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
robots-txt-parser==1.0.0

# Banco Vetorial RAG
 

# LLMs Integration
openai==1.3.7

 

# Google Sheets
 

# GUI
 

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.2

aiohttp==3.9.1
schedule==1.2.0

# API REST
fastapi==0.115.5
uvicorn==0.32.0
PyJWT==2.9.0
httpx==0.27.2
Django==5.1.2
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-django==4.8.0

# Monitoramento e Logging
 

# Segurança
 

# Configuração
pyyaml==6.0.1
